Make entry connection more robust
When we receive an entry list item for an entry the real time data is
still uninitialized. This makes it awkward to deal with for other
processors.
Instead we should only add an entry to the session on the next
real time update for that entry. Other processors can now safely
check if the entry exists in the session and rely on the data
being initialized.