Initialize distance driven for entries
if the entry is unknown to this processor we want to initialize its
distance driven value. If we don't do this, the value will jump from
zero to its current value and will be wrongly corrected.
There is still the chance that the entry is right in the area where
the distance is not correct but that is a very small change.